Move schema validation code from compiler.go to json_schema_validation.go
- Extract all JSON schema validation related functions to dedicated file
- Move httpURLLoader, validateWorkflowSchema*, formatSchemaValidationError, extractErrorMeta, formatValidationMessage, generateValidationHint, and extractContextLines
- Remove unused imports from compiler.go (net/http, jsonschema, mapper)
- Maintain all existing functionality and method signatures
- All tests continue to pass with the refactored code structure
